requirements.txt
Pillow==9.0.1
Pillow==11.1.0
torch==2.5.0
torchvision==0.20.0
tk==8.6
